♻️refactor: 성별 관련 전역변수 적용
